requirements.txt
uvicorn~=0.37.0
fastapi~=0.119.0
pydantic~=2.12.2
pip~=25.2
zlib~=1.3.1
openssl~=3.0.18
cryptography~=46.0.3
pygments~=2.19.2
packaging~=25.0
setuptools~=80.9.0
wheel~=0.45.1
platformdirs~=4.5.0
bcrypt~=5.0.0
python-jose~=3.5.0
sqlalchemy~=2.0.44